Move available labels to redux, add placeholder actions
