+++ src/redux/reducer.ts
@@ -2,12 +2,14 @@ import { combineReducers } from 'redux';
 import image from './reducers/image';
 import options from './reducers/options';
 import undoHistory from './reducers/undoHistory';
+import labels from './reducers/labels';
 
 // Combines state slices for overall app state
 const geoLabelerReducer = combineReducers({
   image,
 	options,
 	undoHistory,
+  labels,
 });
 
 export const labelViewerReducer = combineReducers({
